feat: update bundle script
