+++ demo/src/App.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 
 const defaultValue = "<p>Dear <span data-user-id=\"29\" quill-type=\"mention\" contenteditable=\"false\" class=\"mention-blot select-none\" style=\"color: red;\">@Juice</span>​,</p><p></p><p>Welcome to our store! We are thrilled to have you as a new customer and we look forward to serving you.</p><p></p><p>If you have any <u>questions</u> or need <u>assistance</u>, feel free to reach out to us at any time.</p><p></p><p>Thank you for choosing us!</p><p></p><p><strong>Best regards,</strong></p><p><em>David</em></p><p><a href=\"www.google.com\" rel=\"noopener noreferrer\" target=\"_blank\">www.google.com</a></p>";
 function App() {
-  const [value, setValue] = useState<RFQValue>('');
+  const [value, setValue] = useState<RFQValue>(defaultValue);
   const onChange = (html: string) => {
     setValue(html);
   };
