--- examples/embed-with-cdn/index.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <script crossorigin src="https://unpkg.com/react@18/umd/react.production.min.js"></script>
-    <script crossorigin src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"></script>
-    <script src="https://unpkg.com/@babel/standalone@7.12.9/babel.min.js"></script>
-
-    <!-- <script type="module" src="/dist/index.esm.js"></script> -->
-    <!-- <script src="https://cdn.jsdelivr.net/npm/react-for-quill@latest/dist/index.esm.js"></script> -->
-
-    <link href="https://cdn.jsdelivr.net/npm/react-for-quill@latest/dist/quill.snow.css" rel="stylesheet" />
-</head>
-
-<body>
-    <div id='root'></div>
-    <script type="text/babel">
-        const e = React.createElement;
-        const MyQuillComponent = () => {
-            const handleQuillChange = (value) => {
-                console.log(value);
-            };
-            return e('div', {children: 'hello world'});
-            // return e(window.ReactForQuill, {
-            //     theme: 'snow',
-            //     onChange: handleQuillChange,
-            //     modules: {
-            //         // Modules configuration
-            //     }
-            // });
-        };
-
-        // Render your component
-        ReactDOM.render(e(MyQuillComponent), document.getElementById('root'));
-    </script>
-</body>
-
-</html>
